requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.12
langchain-community==0.0.29
qiskit==1.0.2
qiskit-aer==0.13.3
spacy==3.7.4
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.6.3
SpeechRecognition==3.10.1
gTTS==2.5.1
pytest==8.1.1
httpx==0.27.0
pyaudio==0.2.14
openai>=1.0.0
google-generativeai>=0.5.0
supabase>=2.3.0
